requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
bidict==0.23.1
click==8.2.0
frozenlist==1.6.0
h11==0.16.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
uvicorn==0.34.2
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
